Move compiler builtins to flags sources folder
